requirements.txt
fastapi>=0.112.1,<1
mariadb>=1.1.10,<2
pydantic>=2.8.2,<3
python-dotenv>=1.0.1,<2
pyyaml>=6.0.2,<7
requests>=2.32.3,<3
uvicorn[standard]>=0.30.6,<0.31

